backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-login==0.6.3
authlib==1.4.0
requests==2.32.3
pillow==11.1.0
python-dotenv==1.0.1
flask-cors==5.0.0